+++ CHANGES.md
@@ -62,7 +62,6 @@
 * The `-Dcantaloupe.list_fonts` VM argument has been replaced with the
   `-list-fonts` command-line argument.
 * Added a `local_uri` key to the delegate script `context` hash.
-* JRuby has been updated to version 9.2, which supports Ruby 2.5.
 
 ## 4.1.6
 
@@ -73,6 +72,8 @@
 * Fixed information responses failing to respect the `page` URL query argument.
 * Fixed a bug that caused requests for data beyond an offset of 2^31 from
   S3Source and AzureStorageSource to fail when chunking is enabled.
+* Updated JRuby to version 9.2.11.1, which addresses multiple CVEs.
+* Updated Jackson to version 2.11.0, which addresses multiple CVEs.
 
 ## 4.1.5